create new class for ConnectedThread
