+++ SoundLevelApp/app/src/main/java/ca/concordia/teamc/soundlevelapp/ConnectedThread.java
@@ -1,0 +1,71 @@
+package ca.concordia.teamc.soundlevelapp;
+
+import android.bluetooth.BluetoothSocket;
+import android.os.SystemClock;
+
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
+
+/**
+ * Created by dinaalyousef on 2018-03-11.
+ */
+
+public class ConnectedThread extends Thread {
+    private final BluetoothSocket mmSocket;
+    private final InputStream mmInStream;
+    private final OutputStream mmOutStream;
+
+    public ConnectedThread(BluetoothSocket socket) {
+        mmSocket = socket;
+        InputStream tmpIn = null;
+        OutputStream tmpOut = null;
+
+        // Get the input and output streams, using temp objects because
+        // member streams are final
+        try {
+            tmpIn = socket.getInputStream();
+            tmpOut = socket.getOutputStream();
+        } catch (IOException e) { }
+
+        mmInStream = tmpIn;
+        mmOutStream = tmpOut;
+    }
+
+    public void run() {
+        byte[] buffer = new byte[1024];  // buffer store for the stream
+        int bytes; // bytes returned from read()
+        // Keep listening to the InputStream until an exception occurs
+        while (true) {
+            try {
+                // Read from the InputStream
+                bytes = mmInStream.available();
+                if(bytes != 0) {
+                    SystemClock.sleep(100); //pause and wait for rest of data. Adjust this depending on your sending speed.
+                    bytes = mmInStream.available(); // how many bytes are ready to be read?
+                    bytes = mmInStream.read(buffer, 0, bytes); // record how many bytes we actually read
+                    //handler.obtainMessage(2, bytes, -1, buffer).sendToTarget(); // Send the obtained bytes to the UI activity
+                }
+            } catch (IOException e) {
+                e.printStackTrace();
+
+                break;
+            }
+        }
+    }
+
+    /* Call this from the main activity to send data to the remote device */
+    public void write(String input) {
+        byte[] bytes = input.getBytes();           //converts entered String into bytes
+        try {
+            mmOutStream.write(bytes);
+        } catch (IOException e) { }
+    }
+
+    /* Call this from the main activity to shutdown the connection */
+    public void cancel() {
+        try {
+            mmSocket.close();
+        } catch (IOException e) { }
+    }
+}
