+++ SoundLevelApp/app/src/main/java/ca/concordia/teamc/soundlevelapp/ListDevicesActivity.java
@@ -58,17 +58,9 @@ public class ListDevicesActivity extends AppCompatActivity {
         //set adapter for list view
         bluetoothDevicesList.setAdapter(BTArrayAdapter);
 
-
         bluetoothDevicesList.setOnItemClickListener(devicesClickListener);
 
     }
-
-//    private AdapterView.OnItemClickListener BTDeviceClickListener = new AdapterView.OnItemClickListener() {
-//        public void onItemClick(AdapterView parent, View v, int position, long id) {
-//            Intent intent = new Intent(ListDevicesActivity.this, MeterConfigScreen.class);
-//            startActivity(intent);
-//        }
-//    };
 
     public void listPairedDevices(View V) {
         // clear list
@@ -162,65 +154,6 @@ public class ListDevicesActivity extends AppCompatActivity {
     private BluetoothSocket createBluetoothSocket(BluetoothDevice device) throws IOException {
         return  device.createRfcommSocketToServiceRecord(BTMODULEUUID);
         //creates secure outgoing connection with BT device using UUID
-    }
-
-    private class ConnectedThread extends Thread {
-        private final BluetoothSocket mmSocket;
-        private final InputStream mmInStream;
-        private final OutputStream mmOutStream;
-
-        public ConnectedThread(BluetoothSocket socket) {
-            mmSocket = socket;
-            InputStream tmpIn = null;
-            OutputStream tmpOut = null;
-
-            // Get the input and output streams, using temp objects because
-            // member streams are final
-            try {
-                tmpIn = socket.getInputStream();
-                tmpOut = socket.getOutputStream();
-            } catch (IOException e) { }
-
-            mmInStream = tmpIn;
-            mmOutStream = tmpOut;
-        }
-
-        public void run() {
-            byte[] buffer = new byte[1024];  // buffer store for the stream
-            int bytes; // bytes returned from read()
-            // Keep listening to the InputStream until an exception occurs
-            while (true) {
-                try {
-                    // Read from the InputStream
-                    bytes = mmInStream.available();
-                    if(bytes != 0) {
-                        SystemClock.sleep(100); //pause and wait for rest of data. Adjust this depending on your sending speed.
-                        bytes = mmInStream.available(); // how many bytes are ready to be read?
-                        bytes = mmInStream.read(buffer, 0, bytes); // record how many bytes we actually read
-                        //handler.obtainMessage(2, bytes, -1, buffer).sendToTarget(); // Send the obtained bytes to the UI activity
-                    }
-                } catch (IOException e) {
-                    e.printStackTrace();
-
-                    break;
-                }
-            }
-        }
-
-        /* Call this from the main activity to send data to the remote device */
-        public void write(String input) {
-            byte[] bytes = input.getBytes();           //converts entered String into bytes
-            try {
-                mmOutStream.write(bytes);
-            } catch (IOException e) { }
-        }
-
-        /* Call this from the main activity to shutdown the connection */
-        public void cancel() {
-            try {
-                mmSocket.close();
-            } catch (IOException e) { }
-        }
     }
 
     @Override
